Check/uncheck all sorta works takes 5 seconds
Doesn't toggle all switches and it's too small and it takes forever, but technically it works repeatedly.
It takes so long though that the first few times I ran it, the last 8 countries got .show AND .collapse as classes. it mysteriously stopped doing that.
Did it by adding a class to all elements called .multi-collapse and making that class be the data-target for the main switch